+++ maslov.aleksey/S2/queue.hpp
@@ -2,7 +2,6 @@
 #define QUEUE_HPP
 
 #include <cstddef>
-#include <algorithm>
 #include <stdexcept>
 
 namespace maslov
